yara-verificator/src/requirements.txt
Flask==3.0.0
yara-python==4.3.1
gunicorn==21.2.0